+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     * This is to support the new rule for Hunters, where they were changed from doubling their AGI mod for figuring Charge and Run to instead gain a +3 bonus to their AGI mod when calculating those movement speeds. This field can take a number input, which will allow players to use either the new or old rules, as well as for use in like abilities, both for homebrew content and possible future official content.
   * Updated a number of system terms that have been changed for 5.1.
   * Fixed wound calculation to now be accurate for both Named Characters and Bestiary.
+  * Updated rate of melee attacks.
 * Re-introduced the ammunition management feature, with a new system setting offering the following three modes:
   * Magazines - weapons track full magazines, and reloaded weapons lose the held ammunition in the current magazine.
   * Ammo Pool - a count of total rounds are tracked, with magazines always topped off. This works like how the games handle ammunition, and is how single-loading weapons will work even in Magazines mode.
@@ -36,7 +37,7 @@
 * Weapons have a Scope Magnification value, and display their effective minimum and maximum ranges with that magnification factored in.
 * Attacks that deal Special Damage now indicate this by displaying a bullseye.
 * Equipment Item Compendia are now available.
-* Field added to melee weapons to indicate the number of extra attacks it grants on a Full Attack action.
+* Field added to melee weapons to indicate the number of extra attacks it grants on an Attack action.
 * Field added to thrown weapons to indicate a flat range bonus granted by the weapon.
 
 ## Bug Fixes